Change "a" to "file"
The variable "a" causes issues in Discord especially when you're using it to get server points to get roles. It can cause channels to automatically change, members nicknames to change to your nickname. My nickname was "BIG HARD" and it subsequently caused members nicknames to get changed to my nickname.